logging: move logging wrapper and avoid expensive concatination
